refactor: move file name validation out of controller
And into a new validators module.
This change is to separate the concerns of the controller and the validation logic.
The controller should only be responsible for handling the request and response,
while the validation logic should be in a separate module.

Keeps the controller thin and sets up further engineers with a place to
add more validation logic in the future.