refactor: existing classifier moved to new module
- Moved existing classifier logic to a new module
- Created a base class for file classifiers
- Created a new classifier for heron data files
- Updated the app to use the new classifier

This is to allow the mulitple classifiers to be used in the future to
speed up testing, without having to change the app code nor remove any
existing classifiers.

We may want to use different classifiers for different file types, or
even use multiple classifiers for the same file type to increase
accuracy.