feat: Add Heron Data Classifier v2: a fine-tuned CNN
This is a model that's been fine-tuned using the ResNet18 CNN
as a pre-trained base model.

We need to convert any pdfs into images and then use
the image of the first page of the pdf as the data
to make an inference/prediction on.

Needed to update the integration tests to use the actual
bank_statements, drivers_licenses and invoices, rather
than dummy data - else the model doesn't work. This
is exactly how requests are made to the app so this
is preferable to dummy data.

Have refactored the entire src/services namespace/module
to incorporate this new classifier and to allow versioned
heron data classifiers. This makes the addition of
further classifiers more extendable.